Started implementing ChangeSpell

+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/ChangeSpell.cs
@@ -6,7 +6,7 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
 {
     public class ChangeSpell : BasePacket
     {
-        public ChangeSpell(Unit unit, int slot, string spell)
+        public ChangeSpell(Unit unit, byte slot, string spell)
             : base(PacketCmd.PKT_S2C_ChangeSpell, unit.NetId)
         {
             buffer.Write((byte)slot);
